+++ packages/algolia-search/plugins/AlgoliaSearchResultsPlugin.tsx
@@ -1,23 +1,22 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { ProductItemsGridProps } from '@graphcommerce/magento-product'
 import { IfConfig, PluginProps } from '@graphcommerce/next-config'
-import { useAlgoliaResults } from '../hooks/useAlgoliaResults'
+import dynamic from 'next/dynamic'
 import { useSearchRoute } from '../hooks/useSearchRoute'
 
 export const component = 'ProductListItemsBase'
 export const exported = '@graphcommerce/magento-product'
 export const ifConfig: IfConfig = 'demoMode'
 
-function AlgoliaResults(props: PluginProps<ProductItemsGridProps>) {
-  const { Prev } = props
-  const { products } = useAlgoliaResults()
-  return <Prev {...props} items={products} />
-}
-
 function AlgoliaSearchPlugin(props: PluginProps<ProductItemsGridProps>) {
   const { Prev } = props
   const search = useSearchRoute()
+
   if (!search) return <Prev {...props} />
+
+  const AlgoliaResults = dynamic(
+    async () => (await import('../components/Filters/AlgoliaResults')).AlgoliaResults,
+  )
 
   return <AlgoliaResults {...props} />
 }
